# src/back/eco_farma_API/eco_farma_API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "eco_farma_API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Entrega/{id}": {
      "get": {
        "tags": [
          "Entrega"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Entrega"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entrega"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Entrega"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Entrega"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Entregador": {
      "get": {
        "tags": [
          "Entregador"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Entregador"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Entregador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Entregador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Entregador"
            }
          }
        }
      }
    },
    "/api/Entregador/{id}": {
      "get": {
        "tags": [
          "Entregador"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Entregador"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Entregador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Entregador"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Entregador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Farmacia": {
      "get": {
        "tags": [
          "Farmacia"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Farmacia"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Farmacia"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Farmacia"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Farmacia"
            }
          }
        }
      }
    },
    "/api/Farmacia/{id}": {
      "get": {
        "tags": [
          "Farmacia"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Farmacia"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Farmacia"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Farmacia"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Farmacia"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pedido": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pedido"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pedido"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        }
      }
    },
    "/api/Pedido/{id}": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pedido"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pedido"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Entrega": {
      "type": "object",
      "properties": {
        "id_entrega": {
          "format": "int32",
          "type": "integer"
        },
        "id_pedido": {
          "format": "int32",
          "type": "integer"
        },
        "id_entregador": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Entregador": {
      "type": "object",
      "properties": {
        "id_entregador": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "sexo": {
          "type": "string"
        },
        "data_nasc": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "cpf": {
          "type": "string"
        },
        "endereco": {
          "type": "string"
        },
        "cep": {
          "type": "string"
        },
        "numero": {
          "format": "int32",
          "type": "integer"
        },
        "senha": {
          "type": "string"
        }
      }
    },
    "Farmacia": {
      "type": "object",
      "properties": {
        "id_farmacia": {
          "format": "int32",
          "type": "integer"
        },
        "numero": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "cnpj": {
          "type": "string"
        },
        "telefone": {
          "type": "string"
        },
        "cep": {
          "type": "string"
        },
        "endereco": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        }
      }
    },
    "Pedido": {
      "type": "object",
      "properties": {
        "id_pedido": {
          "format": "int32",
          "type": "integer"
        },
        "id_cliente": {
          "format": "int32",
          "type": "integer"
        },
        "id_farmacia": {
          "format": "int32",
          "type": "integer"
        },
        "id_produto": {
          "format": "int32",
          "type": "integer"
        },
        "qtd_produto": {
          "format": "int32",
          "type": "integer"
        },
        "preco_produto": {
          "format": "double",
          "type": "number"
        }
      }
    },
...
  }
}